requirements.txt
feedparser>=6.0
beautifulsoup4>=4.12
lxml>=4.9
requests>=2.28
PyYAML>=6.0
flask>=3.0
flask-cors>=4.0
jieba>=0.42
